@@ -1,34 +1,31 @@
 {
   "name": "millegriles.collection-client",
-  "version": "2022.4.0",
+  "version": "2022.5.0",
   "private": true,
   "proxy": "https://localhost:3037",
   "homepage": "/collections",
   "dependencies": {
     "@babel/plugin-syntax-object-rest-spread": "^7.8.3",
+    "buffer": "^6.0.3",
     "comlink": "^4.3.1",
+    "crypto-browserify": "^3.12.0",
     "font-awesome": "^4.7.0",
-    "react-scripts": "^4",
-    "react-transition-group": "^4.4.2",
-    "rescript-worker-loader": "^1.0.4",
-    "wasm-loader": "^1.3.0",
-    "worker-loader": "^3.0.8"
+    "path-browserify": "^1.0.1",
+    "react-app-rewired": "^2.2.1",
+    "react-scripts": "^5",
+    "react-transition-group": "^4.4.5",
+    "stream-browserify": "^3.0.0"
   },
   "peerDependencies": {
-    "@dugrema/millegrilles.reactjs": "git+ssh://git.maceroc.com/git/millegrilles.reactjs#2022.4",
+    "@dugrema/millegrilles.reactjs": "git+ssh://git.maceroc.com/git/millegrilles.reactjs#2022.5",
     "bootstrap": "^5.1",
-    "react": "^17",
+    "react": "^18",
     "react-bootstrap": "^2.4.0",
-    "react-dom": "^17",
-    "webpack": "~4.44.2"
-  },
-  "devDependencies": {
-    "@rescripts/cli": "^0.0.16"
+    "react-dom": "^18"
   },
   "scripts": {
-    "start": "export WDS_SOCKET_PATH=/collections/sockjs-node && rescripts start",
-    "build": "rescripts build",
-    "test": "react-scripts test",
+    "start": "react-app-rewired start",
+    "build": "BABEL_ENV=production react-app-rewired build",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
